Created demo dir, moved demo_pm to separate file.
