+++ demo/README.md
@@ -1,0 +1,9 @@
+This directory contains procedures that demonstrate some of kernel
+features. They are meant to serve as functionality proofs and
+examples, and therefore are a valuable documentation.
+
+When designing and/or implementing a new interface, you are welcome to
+drop an example usage in this directory.
+
+These files are not compiled by default, if you wish to run a
+demonstration, you need to compile and call it on your own.
