+++ demo/demos.h
@@ -1,0 +1,10 @@
+#ifndef __DEMOS_H__
+#define __DEMOS_H__
+
+/* A trivial demonstration of physical memory manager features:
+ * Allocating a few different-sized batches of frames and freeing them
+ * aferwards, printing manager's internal state after each
+ * operation. */
+void demo_pm();
+
+#endif // __DEMOS_H__
